lab_20_py/requirements.txt
transformers
torch
Pillow

